move product manager to file
