+++ productManager.js
@@ -1,0 +1,32 @@
+class ProductManager {
+  constructor() {
+    this.products = [];
+  }
+
+  addProduct = (title, description, price, thumbnail, code, stock) => {
+    const newProduct = {
+      id: this.products.length + 1,
+      title,
+      description,
+      price,
+      thumbnail,
+      code,
+      stock,
+    };
+    this.products.push(newProduct);
+    return newProduct;
+  };
+
+  getProducts() {
+    return this.products;
+  }
+
+  getProductById(id) {
+    const product = this.products.find((product) => product.id === id);
+    if (!product) {
+      throw new Error("Not found");
+    }
+    return product;
+  }
+}
+module.exports = ProductManager;
